Alter tap-dance to always trigger behaviors on every key press.
This is a much better behavior when using tap-dance with modifiers,
since it's okay to press them needlessly, but it sucks when they don't
trigger immediately.  The previous behavior worked fine just for typing,
since it gets interrupted and immediately triggers upon another key
press happening.  But modifer+mouse combos, it introduced a noticeable
delay that caused problems (e.g. when trying to do viewport navigation
in Blender).

In my case, I'm *only* using tap-dance with modifiers, so I'm just
changing it's behavior entirely.  If I eventually want to upstream
this, it will need to be an option on the behavior.